+++ client/src/modules/chat-message.js
@@ -4,7 +4,7 @@ const addMessageToLogs = require('./add-message-to-logs');
 
 // Multi-player chat.
 const chatMessage = () => {
-  socket.on('chatMessage', (chatMessage) => addMessageToLogs(chatMessage));
+  EG.socket.on('chatMessage', (chatMessage) => addMessageToLogs(chatMessage));
 
   const logsForm = document.getElementById('logsForm');
 
@@ -13,7 +13,7 @@ const chatMessage = () => {
     const chatMessage = document.getElementById('logsInput').value;
 
     // Emit current player chat message to all players, via server side.
-    socket.emit('chatMessage', { chatMessage, name: EG.dataStore.player.name });
+    EG.socket.emit('chatMessage', { chatMessage, name: EG.dataStore.player.name });
 
     // Clear content and remove focus from logs input field.
     const logsInput = document.getElementById('logsInput');