+++ server/spawn-player.js
@@ -1,6 +1,6 @@
 'strict';
 
 // Spawn current player.
-const spawnPlayer = (io, socketId, player) => io.to(socketId).emit('spawnPlayer', player);
+const spawnPlayer = (io, socketId, player) => EG.io.to(socketId).emit('spawnPlayer', player);
 
 module.exports = spawnPlayer;