+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Move sound class to be a reusable, external dependency.
 - Message all other players when a player died.
 - Set master volume for Web Audio API sounds.
+- npc respawn at their respawn position once their respawn timer has been hit since they died.
 - npc no longer fight npc.
 - npc move only on client side, with a small animation, but not from server-side, to lighten the load.